+++ README.md
@@ -1,32 +1,13 @@
-Rewards Restful
-===============================
-
-> WebAPI / Spring
-
- 
-
-A retailer offers a rewards program to its customers, awarding points based on each recorded purchase.
-
- 
-
-A customer receives 2 points for every dollar spent over $100 in each transaction, plus 1 point for every dollar spent over $50 in each transaction
-
-(e.g. a $120 purchase = 2x$20 + 1x$50 = 90 points).
-
- 
-
-Given a record of every transaction during a three month period, calculate the reward points earned for each customer per month and total.
-
-* Make up a data set to best demonstrate your solution
-* Check solution into GitHub
 
 
-  
 
-------------------------------------------------------------
-
-
-###To Run this project locally
+###To Run this project, do Run command and test with a Postman collection
 ```shell
 $ git clone https://github.com/cylixx/rewards-restful.git
 $ mvn spring-boot: run
+
+#getCustomerby Id:
+###http://localhost:8080/customers/102
+
+#getCustomer
+#http://localhost:8080/customers
