+++ src/main/java/com/company/rewardscalculator/Model/Reward.java
@@ -1,4 +1,4 @@
-package com.example.service.springbootrewards.model;
+package com.company.rewardscalculator.Model;
 
 import javax.persistence.Transient;
 
@@ -6,8 +6,8 @@ import com.fasterxml.jackson.annotation.JsonInclude;
 
 public abstract class Reward {
 
-	@JsonInclude  //include in Jackson serialization
-	@Transient    //marking a field as not persistable
+	@JsonInclude  //Jackson serialization
+	@Transient    //making this field not persistant
 	protected Long points;
 
 	public abstract Long getPoints();
