+++ src/main/java/com/company/rewardscalculator/Service/CustomerRepo.java
@@ -1,0 +1,10 @@
+package com.company.rewardscalculator.Service;
+
+import com.company.rewardscalculator.Model.Customer;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface CustomerRepo extends JpaRepository<Customer, Integer>{
+
+}
