--- src/main/java/com/example/service/springbootrewards/rewards/RewardsService.java
@@ -1,32 +1,0 @@
-package com.example.service.springbootrewards.rewards;
-
-import java.util.List;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.stereotype.Service;
-
-import com.example.service.springbootrewards.model.Customer;
-
-@Service
-public class RewardsService {
-	
-	@Autowired
-	private CustomerRepository customerRepository;
-	
-	
-	public List<Customer> getCustomerAll() {
-		return customerRepository.findAll();
-	}
-	
-	public Customer getCustomerById(Integer customerId) {
-		return customerRepository.findById(customerId).orElse(null);
-	}
-	
-//	public ResponseEntity<Customer> getCustomerById(Integer customerId) {
-//		Customer customer = customerRepository.findById(customerId).orElse(null);
-//		return new ResponseEntity<Customer>(customer, HttpStatus.NOT_FOUND);
-//	}
-
-}
