--- src/main/java/com/example/service/springbootrewards/utils/Utils.java
@@ -1,6 +1,0 @@
-package com.example.service.springbootrewards.utils;
-
-public interface Utils {
-
-	//public final abstract REWARD_PERCENTAGE;
-}
