+++ src/main/resources/application.properties
@@ -1,7 +1,6 @@
 spring.jpa.show-sql=true
 spring.h2.console.enabled=true
-#H2L: http://localhost:8080/h2-console
-#JDBC URL:  jdbc:h2:mem:testdb   <-- connect with data base
+
